requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
execnet==2.0.2
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
PyJWT==2.8.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-json-report==1.5.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-timeout==2.1.0
pytest-xdist==3.3.1
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
tomli==2.0.1
twilio==8.5.0
urllib3==2.0.4
yarl==1.9.2
